+++ src/components/Form/Col/Col.tsx
@@ -17,8 +17,8 @@ import FormFieldNumberTable from "@root/components/Field/FormFieldNumberTable";
 import FormFieldPhoneSelectionDropdown from "@root/components/Field/FormFieldPhoneSelectionDropdown";
 import FormFieldRadio from "@root/components/Field/FormFieldRadio";
 import FormFieldRaw from "@root/components/Field/FormFieldRaw";
+import FormFieldTable from "@root/components/Field/FormFieldTable";
 import FormFieldToggleSwitch from "@root/forms/FormFieldToggleSwitch";
-import FormFieldTable from "@root/forms/FormFieldTable";
 import FormFieldText from "@root/forms/FormFieldText";
 import FormFieldTextEditor from "@root/forms/FormFieldTextEditor";
 import FormFieldUpload from "@root/forms/FormFieldUpload";
